% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{apply_function_to_column}
\alias{apply_function_to_column}
\title{Apply function to multiple columns.}
\usage{
apply_function_to_column(frame, list_of_columns, new_col, funct)
}
\arguments{
\item{data}{frame}

\item{list}{of columns to apply function to}

\item{name}{of new column}

\item{function}{to apply}
}
\value{
original data frame with new column attached
}
\description{
This function enables the application of a function to multiple columns.
}
