% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{encode_and_bind}
\alias{encode_and_bind}
\title{One-hot encode categorical variables.}
\usage{
encode_and_bind(frame, feature_to_encode)
}
\arguments{
\item{data}{frame}

\item{feature}{to encode}
}
\value{
hot-encoded data frame
}
\description{
This function one-hot encodes categorical variables and binds the entire frame.
}
