% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_closest_string}
\alias{get_closest_string}
\title{Discover closest matching string from set.}
\usage{
get_closest_string(vector_of_strings, search_string)
}
\arguments{
\item{vector}{of strings}

\item{search}{string}
}
\value{
closest matching string
}
\description{
This function discovers the closest matching string from vector of strings.
}
