% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{remove_features}
\alias{remove_features}
\title{Remove specified features from data frame.}
\usage{
remove_features(frame, features)
}
\arguments{
\item{features}{to remove}

\item{data}{frame}
}
\value{
original data frame less removed features
}
\description{
This function removes specified features from data frame.
}
